updating

+++ model/grenc_trdec_model/model.py
@@ -97,7 +97,7 @@ class ClipModel(nn.Module):
         
         else:
             # Train Dec
-            vit_enc_output = vit_enc_output.reshape(vit_enc_output.shape[0],-1) # (B, w*h)
+            vit_enc_output = vit_enc_output.reshape(vit_enc_output.shape[0],-1).long() # (B, w*h)
             embed_fv = self.embed_text(vit_enc_output)  # (b, l, emb)
             xfmer_enc_output = self.Xfmer_ENC(embed_fv).permute(1,0,2) # (B, l, emb_dim)
             xfmer_dec_output = self.Xfmer_DEC(mml,
